WCDH: Fix the C_STANDARD property in the tests.
The CXX_STANDARD property validates the values allowed for it,
and '90' is not a valid value for it.  However, the validation is only
done for languages enabled for the target, and as C_undefined has no
CXX files, that language is not enabled for it.

This amends commit v3.1.0-rc3~23^2 (WCDH: Make the header -Wundef
safe for the C language., 2014-11-24).